Split thermocouple data by location
In the experiment, 7 thermocouples were placed along the inside of the pipe and 6 thermocouples were placed in pairs outside of the pipe. One of the pair was directly outside the pipe, and the other was placed outside the insulation that wrapped the pipe. 3 of these pairs were placed along the length of the apparatus.